// src/levels/level9/types/finish-event-enums.ts
import { FinishEventMapWithDefaults, ObjectiveType } from '@/types/objective-types';

export enum EdgeConnectionFinishEvent {
  RDS1_MANAGE_POLICY_ATTACHED_GROUP1 = 'RDS1_MANAGE_POLICY_ATTACHED_GROUP1',
  RDS2_MANAGE_POLICY_ATTACHED_GROUP2 = 'RDS2_MANAGE_POLICY_ATTACHED_GROUP2',
  RDS_SHARED_MANAGE_POLICY_ATTACHED_GROUP1 = 'RDS_SHARED_MANAGE_POLICY_ATTACHED_GROUP1',
  RDS_SHARED_MANAGE_POLICY_ATTACHED_GROUP2 = 'RDS_SHARED_MANAGE_POLICY_ATTACHED_GROUP2',
}

export enum PolicyCreationFinishEvent {
  RDS1_MANAGE_POLICY_CREATED = 'RDS_MANAGE_POLICY1_CREATED',
  RDS2_MANAGE_POLICY_CREATED = 'RDS2_MANAGE_POLICY1_CREATED',
  RDS_SHARED_POLICY_CREATED = 'RDS_SHARED_POLICY_CREATED',
}

export type FinishEventMap = FinishEventMapWithDefaults<{
  [ObjectiveType.POLICY_CREATION_OBJECTIVE]: PolicyCreationFinishEvent;
  [ObjectiveType.EDGE_CONNECTION_OBJECTIVE]: EdgeConnectionFinishEvent;
}>;